:zap: move CaseInsensitiveCompare into utils
